move back raw mapping
